# docs/swagger.yaml
basePath: /
definitions:
...
  model.District:
    description: İlçe bilgileri
    properties:
      name:
        description: İlçe adı
        example: Beşiktaş
        type: string
      province:
        allOf:
        - $ref: '#/definitions/model.Province'
        description: İl bilgisi
      province_id:
        description: Bağlı olduğu il
        example: 1
        type: integer
    required:
    - name
    - province_id
    type: object
  model.Hospital:
    description: Hastane bilgileri
    properties:
      address_detail:
        description: Açık adres
        example: Beşiktaş Caddesi No:123
        type: string
      district:
        allOf:
        - $ref: '#/definitions/model.District'
        description: İlçe bilgisi
      district_id:
        description: İlçe ID
        example: 1
        type: integer
      email:
        description: E-posta adresi
        example: info@acibadem.com
        type: string
      name:
        description: Hastane adı
        example: Acıbadem Hastanesi
        type: string
      phone:
        description: Telefon numarası
        example: "02121234567"
        type: string
      province:
        allOf:
        - $ref: '#/definitions/model.Province'
        description: İlişkiler
      province_id:
        description: İl ID
        example: 1
        type: integer
      tax_id:
        description: Vergi kimlik numarası
        example: "1234567890"
        type: string
      users:
        description: Hastane kullanıcıları
        items:
          $ref: '#/definitions/model.User'
        type: array
    required:
    - address_detail
    - district_id
    - email
    - name
    - phone
    - province_id
    - tax_id
    type: object
...
  model.HospitalRegistrationRequest:
    description: Hastane kayıt verisi
    properties:
      address_detail:
        example: Beşiktaş Caddesi No:123
        type: string
      admin_email:
        example: ahmet.yilmaz@acibadem.com
        type: string
      admin_first_name:
        description: Yetkili Bilgileri
        example: Ahmet
        type: string
      admin_last_name:
        example: Yılmaz
        type: string
      admin_password:
        example: "123456"
        minLength: 6
        type: string
      admin_phone:
        example: "05551234567"
        type: string
      admin_tc:
        example: "12345678901"
        type: string
      district_id:
        example: 1
        type: integer
      hospital_email:
        example: info@acibadem.com
        type: string
      hospital_name:
        description: Hastane Bilgileri
        example: Acıbadem Hastanesi
        type: string
      hospital_phone:
        example: "02121234567"
        type: string
      province_id:
        description: Adres Bilgileri
        example: 1
        type: integer
      tax_id:
        example: "1234567890"
        type: string
    required:
    - address_detail
    - admin_email
    - admin_first_name
    - admin_last_name
    - admin_password
    - admin_phone
    - admin_tc
    - district_id
    - hospital_email
    - hospital_name
    - hospital_phone
    - province_id
    - tax_id
    type: object
  model.HospitalRegistrationResponse:
    description: Hastane kayıt yanıtı
    properties:
      admin_user:
        $ref: '#/definitions/model.User'
      hospital:
        $ref: '#/definitions/model.Hospital'
      message:
        example: Hastane başarıyla kaydedildi
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  model.LoginRequest:
    description: Kullanıcı giriş bilgileri
    properties:
      email:
        description: E-posta adresi
        example: ahmet.yilmaz@example.com
        type: string
      password:
        description: Şifre
        example: "123456"
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
  model.Polyclinic:
    description: Hastane poliklinik bilgileri (eski model)
    properties:
      floor:
        description: Kat numarası
        example: 2
        type: integer
      name:
        description: Poliklinik adı
        example: Kardiyoloji
        type: string
      room_number:
        description: Oda numarası
        example: 205
        type: integer
    required:
    - floor
    - name
    - room_number
    type: object
...
  model.Province:
    description: İl bilgileri
    properties:
      districts:
        description: İlçeler
        items:
          $ref: '#/definitions/model.District'
        type: array
      name:
        description: İl adı
        example: İstanbul
        type: string
    required:
    - name
    type: object
...
  model.User:
    description: Hastane kullanıcı bilgileri
    properties:
      created_by:
        description: Kim tarafından eklendi (nullable - ilk user için)
        example: 1
        type: integer
      creator:
        allOf:
        - $ref: '#/definitions/model.User'
        description: Oluşturan kullanıcı
      email:
        description: E-posta adresi
        example: ahmet.yilmaz@example.com
        type: string
      first_name:
        description: Ad
        example: Ahmet
        type: string
      hospital:
        allOf:
        - $ref: '#/definitions/model.Hospital'
        description: İlişkiler
      hospital_id:
        description: Hangi hastaneye ait
        example: 1
        type: integer
      is_active:
        description: Aktif mi?
        example: true
        type: boolean
      last_name:
        description: Soyad
        example: Yılmaz
        type: string
      password:
        description: Şifre (JSON'dan okuyabilir ama response'da göstermez)
        example: "123456"
        minLength: 6
        type: string
      phone:
        description: Telefon numarası
        example: "05551234567"
        type: string
      role:
        description: 'Rol: "yetkili" veya "çalışan"'
        enum:
        - yetkili
        - çalışan
        example: çalışan
        type: string
      tc:
        description: Türkiye Cumhuriyeti Kimlik Numarası
        example: "12345678901"
        type: string
    required:
    - email
    - first_name
    - hospital_id
    - last_name
    - password
    - phone
    - role
    - tc
    type: object
host: localhost:8080
info:
  contact: {}
  description: Bu API VatanSoft staj giriş projesi için yazılmıştır.
  title: Hastane Takip API
  version: "1.0"
paths:
  /hospital/{id}:
    get:
      description: ID'ye göre hastane bilgilerini getirir
      parameters:
      - description: Hastane ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Hospital'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Hastane bilgilerini getir
      tags:
      - Hospital
...
  /hospital/register:
    post:
      consumes:
      - application/json
      description: Yeni hastane ve admin kullanıcı kaydı yapar
      parameters:
      - description: Hastane kayıt verisi
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.HospitalRegistrationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.HospitalRegistrationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Hastane kaydı
      tags:
      - Hospital
...
  /login:
    post:
      consumes:
      - application/json
      description: Email ve şifre ile kullanıcı girişi yapılır
      parameters:
      - description: Giriş bilgileri
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Kullanıcı girişi
      tags:
      - Auth
...
  /polyclinics:
    get:
      description: Sistemdeki tüm poliklinikleri listeler.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Polyclinic'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Tüm poliklinikleri getir
      tags:
      - Polyclinic
    post:
      consumes:
      - application/json
      description: Yeni bir poliklinik ekler.
      parameters:
      - description: Poliklinik verisi
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.Polyclinic'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Yeni poliklinik oluştur
      tags:
      - Polyclinic
  /polyclinics/{id}:
    delete:
      description: ID'ye göre poliklinik kaydını siler.
      parameters:
      - description: Poliklinik ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Poliklinik sil
      tags:
      - Polyclinic
    put:
      consumes:
      - application/json
      parameters:
      - description: Poliklinik ID
        in: path
        name: id
        required: true
        type: integer
      - description: Yeni poliklinik verisi
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.Polyclinic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Poliklinik bilgilerini güncelle
      tags:
      - Polyclinic
  /provinces:
    get:
      description: Dropdown için tüm illeri listeler
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Province'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Tüm illeri getir
      tags:
      - Location
  /provinces/{province_id}/districts:
    get:
      description: Seçilen ile ait ilçeleri listeler
      parameters:
      - description: İl ID
        in: path
        name: province_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.District'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: İle göre ilçeleri getir
      tags:
      - Location
  /register:
    post:
      consumes:
      - application/json
      description: Yeni kullanıcıyı kaydeder
      parameters:
      - description: Kullanıcı bilgileri
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Yeni kullanıcı kaydı
      tags:
      - Auth
...
swagger: "2.0"